Completed transition to economics. Added a complete test.
