+++ CombiCSP/__init__.py
@@ -9,4 +9,4 @@ from CombiCSP.misc import CtoK, OutputContainer
 from CombiCSP.solar_system_location import SolarSystemLocation
 from CombiCSP.solar_tower import SolarTowerCalcs
 from CombiCSP.solar_trough import SolarTroughCalcs
-
+from CombiCSP.economics import Economic_environment
